// types/transactionCard.ts
export interface LastFiveTransactionprops {
  TRANSACTION_NUMBER: string;
  PROCEDURE_ID: string;
  GROUP_ID: string;
  GROUP_AR: string;
  GROUP_EN: string;
  PROCEDURE_AR: string;
  PROCEDURE_EN: string;
  INSTANCE_DATE: { value: string };
  PROPERTY_TYPE_ID: string;
  PROP_TYPE_AR: string;
  PROP_TYPE_EN: string;
  PROPERTY_SUB_TYPE_ID: string;
  PROP_SB_TYPE_AR: string;
  PROP_SB_TYPE_EN: string;
  USAGE_AR: string;
  USAGE_EN: string;
  IS_OFFPLAN: string;
  IS_OFFPLAN_AR: string;
  IS_OFFPLAN_EN: string;
  AREA_ID: string;
  AREA_AR: string;
  AREA_EN: string;
  building_name_ar: string | null;
  building_name_en: string | null;
  project_number: string;
  PROJECT_AR: string;
  PROJECT_EN: string;
  master_project_en: string | null;
  master_project_ar: string | null;
  nearest_landmark_ar: string | null;
  nearest_landmark_en: string | null;
  nearest_metro_ar: string | null;
  nearest_metro_en: string | null;
  nearest_mall_ar: string | null;
  nearest_mall_en: string | null;
  ROOMS_AR: string;
  ROOMS_EN: string;
  has_parking: string;
  ACTUAL_AREA: string;
  TRANS_VALUE: string;
  meter_sale_price: string;
  rent_value: string | null;
  meter_rent_price: string | null;
  no_of_parties_role_1: string;
  no_of_parties_role_2: string;
  no_of_parties_role_3: string;
  IS_FREE_HOLD: string | null;
  IS_FREE_HOLD_EN: string | null;
  IS_FREE_HOLD_AR: string | null;
}